requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
python-dotenv==1.0.0
requests>=2.31.0
dnspython>=2.4.2
beautifulsoup4>=4.12.2
cachetools>=5.3.1
python-whois>=0.8.0
reportlab>=4.0.4
Werkzeug>=3.0.0
scikit-learn>=1.5.0
networkx>=3.1
plotly>=5.15.0
redis>=5.0.0
schedule>=1.2.0
pycryptodome>=3.18.0
APScheduler>=3.10.4
joblib>=1.3.2
cryptography>=41.0.3
Jinja2>=3.1.2
flask-restful>=0.3.10
flask-limiter>=3.5.0
flask-cors>=4.0.0
PyJWT>=2.8.0
numpy>=1.26.0
pandas>=2.1.0
lxml>=4.9.3
urllib3>=2.0.4
charset-normalizer>=3.2.0
idna>=3.4
certifi>=2023.7.22
PyYAML>=6.0.1
simple-websocket>=1.0.0
python-engineio>=4.8.0
python-socketio>=5.10.0